gRCP-server/requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
Werkzeug==3.1.3
grpcio==1.70.0
grpcio-tools==1.70.0
protobuf==5.29.3